@@ -35,12 +35,21 @@
     "style-loader": "^0.13.1",
     "text-loader": "0.0.1",
     "ts-loader": "^0.8.2",
-    "typescript": "^1.8.10",
-    "webpack": "^1.13.1",
+    "typescript": "^2.1.5",
+    "webpack": "^1.14.0",
     "webpack-validator": "^2.2.0"
   },
   "dependencies": {
-    "bootstrap": "^3.3.6",
+    "@types/i18next": "^2.3.35",
+    "@types/jquery": "^2.0.39",
+    "@types/knockout": "^3.4.39",
+    "@types/knockout.mapping": "^2.0.28",
+    "@types/pubsub-js": "^1.5.17",
+    "@types/sharepoint": "^2013.1.1",
+    "@types/sprintf-js": "0.0.27",
+    "@types/trunk8": "0.0.27",
+    "@types/whatwg-fetch": "0.0.33",
+    "bootstrap": "^3.3.7",
     "es6-promise": "^3.2.1",
     "flickity": "^2.0.3",
     "i18next": "^3.3.1",
@@ -50,8 +59,8 @@
     "moment": "^2.14.1",
     "node-fetch": "^1.5.3",
     "office-ui-fabric": "^2.6.1",
-    "pubsub-js": "^1.5.3",
-    "sp-pnp-js": "1.0.4",
+    "pubsub-js": "^1.5.4",
+    "sp-pnp-js": "^2.0.0",
     "sprintf-js": "^1.0.3",
     "trunk8": "0.0.1",
     "whatwg-fetch": "^1.0.0"
